+++ templar_test.go
@@ -1,12 +1,13 @@
 package templar
 
 import (
-	"os"
-	"testing"
+  "fmt"
+  "os"
+  "testing"
 )
 
-var noDotEnvExpectation = `Hello Oedipus!
-How do you like it in "/Users/runeimp/dev/apps/templar"?
+var exampleDotEnvExpectation = `Hello Oedipus!
+How do you like it in "/Users/runeimp/Dropbox/Profile/Home/dev/apps/templar"?
 
   ENV_FILE_COMMENT == ''
       ENV_FILE_VAR == 'The Bard'
@@ -23,7 +24,7 @@ DEFAULT.global_ini ==
 `
 
 var dotEnvExpectation = `Hello Horatio!
-How do you like it in "/Users/runeimp/dev/apps/templar"?
+How do you like it in "/Users/runeimp/Dropbox/Profile/Home/dev/apps/templar"?
 
   ENV_FILE_COMMENT == ''
       ENV_FILE_VAR == '.env Ninja!'
@@ -40,7 +41,7 @@ DEFAULT.global_ini ==
 `
 
 var dotEnvAndINIExpectation = `Hello Hamlet!
-How do you like it in "/Users/runeimp/dev/apps/templar"?
+How do you like it in "/Users/runeimp/Dropbox/Profile/Home/dev/apps/templar"?
 
   ENV_FILE_COMMENT == ''
       ENV_FILE_VAR == '.env Ninja!'
@@ -57,7 +58,7 @@ DEFAULT.global_ini == true
 `
 
 var dotEnvAndJSONExpectation = `Hello Horatio!
-How do you like it in "/Users/runeimp/dev/apps/templar"?
+How do you like it in "/Users/runeimp/Dropbox/Profile/Home/dev/apps/templar"?
 
   ENV_FILE_COMMENT == ''
       ENV_FILE_VAR == '.env Ninja!'
@@ -73,86 +74,53 @@ DEFAULT.global_ini ==
 
 `
 
+var noDotEnvExpectation = `Hello runeimp!
+How do you like it in "/Users/runeimp/Dropbox/Profile/Home/dev/apps/templar"?
+
+  ENV_FILE_COMMENT == ''
+      ENV_FILE_VAR == ''
+       CLI_ENV_VAR == 'Sound and fury'
+           CLI_VAR == 'As you like it'
+           boolean == 
+     one.two.three == 
+       numbers.two == 
+               all ==
+         words.all == 
+             POSIX == 
+DEFAULT.global_ini == 
+
+`
+
 func TestTemplar(t *testing.T) {
-	tests := []struct {
-		name        string
-		checkDotEnv bool
-		data        []string
-		template    string
-		want        string
-	}{
-		{name: ".env", template: "example.tmpl", checkDotEnv: true, want: dotEnvExpectation},
-		{name: ".env and example.ini", template: "example.tmpl", checkDotEnv: true, data: []string{"example.ini"}, want: dotEnvAndINIExpectation},
-		{name: ".env and example.json", template: "example.tmpl", checkDotEnv: true, data: []string{"example.json"}, want: dotEnvAndJSONExpectation},
-		{name: "example.env", template: "example.tmpl", checkDotEnv: false, data: []string{"example.env"}, want: noDotEnvExpectation},
-	}
+  tests := []struct {
+    name        string
+    checkDotEnv bool
+    data        []string
+    template    string
+    want        string
+  }{
+    {name: ".env", template: "example.tmpl", checkDotEnv: true, want: dotEnvExpectation},
+    {name: ".env and example.ini", template: "example.tmpl", checkDotEnv: true, data: []string{"example.ini"}, want: dotEnvAndINIExpectation},
+    {name: ".env and example.json", template: "example.tmpl", checkDotEnv: true, data: []string{"example.json"}, want: dotEnvAndJSONExpectation},
+    {name: "example.env", template: "example.tmpl", checkDotEnv: false, data: []string{"example.env"}, want: exampleDotEnvExpectation},
+    {name: "no.env", template: "example.tmpl", checkDotEnv: false, want: noDotEnvExpectation},
+  }
 
-	for _, tc := range tests {
-		Reinitialize()
-		os.Setenv("CLI_ENV_VAR", "Sound and fury")
-		os.Setenv("CLI_VAR", "As you like it")
-		if len(tc.data) == 0 {
-			InitData(tc.checkDotEnv)
-		} else {
-			for _, file := range tc.data {
-				InitData(tc.checkDotEnv, file)
-			}
-		}
-		got, _ := Render(tc.template)
-		if tc.want != got {
-			t.Fatalf(`"%s": expected: %v, got: %v`, tc.name, tc.want, got)
-		}
-	}
+  for _, tc := range tests {
+    debug := DebugWarn
+    Reinitialize(debug)
+    os.Setenv("CLI_ENV_VAR", "Sound and fury")
+    os.Setenv("CLI_VAR", "As you like it")
+    if len(tc.data) == 0 {
+      InitData(tc.checkDotEnv)
+    } else {
+      for _, file := range tc.data {
+        InitData(tc.checkDotEnv, file)
+      }
+    }
+    got, _ := Render(tc.template)
+    if tc.want != got {
+      t.Fatalf(fmt.Sprintf("%q:\n\texpected: %v\n\tgot: %v\n\t | tc.checkDotEnv = %t\n", tc.name, tc.want, got, tc.checkDotEnv))
+    }
+  }
 }
-
-// func TestDotEnv(t *testing.T) {
-// 	Reinitialize()
-// 	os.Setenv("CLI_ENV_VAR", "Sound and fury")
-// 	os.Setenv("CLI_VAR", "As you like it")
-// 	checkDotEnv := true
-// 	want := dotEnvExpectation
-// 	InitData(checkDotEnv)
-// 	got, _ := Render("example.tmpl")
-// 	if got != want {
-// 		t.Fatalf("expected: %v, got: %v", want, got)
-// 	}
-// }
-
-// func TestDotEnvAndINI(t *testing.T) {
-// 	Reinitialize()
-// 	os.Setenv("CLI_ENV_VAR", "Sound and fury")
-// 	os.Setenv("CLI_VAR", "As you like it")
-// 	checkDotEnv := true
-// 	want := dotEnvAndINIExpectation
-// 	InitData(checkDotEnv, "example.ini")
-// 	got, _ := Render("example.tmpl")
-// 	if got != want {
-// 		t.Fatalf("expected: %v, got: %v", want, got)
-// 	}
-// }
-
-// func TestDotEnvAndJSON(t *testing.T) {
-// 	Reinitialize()
-// 	os.Setenv("CLI_ENV_VAR", "Sound and fury")
-// 	os.Setenv("CLI_VAR", "As you like it")
-// 	checkDotEnv := true
-// 	want := dotEnvAndJSONExpectation
-// 	InitData(checkDotEnv, "example.json")
-// 	got, _ := Render("example.tmpl")
-// 	if got != want {
-// 		t.Fatalf("expected: %v, got: %v", want, got)
-// 	}
-// }
-
-// func TestNoDotEnv(t *testing.T) {
-// 	Reinitialize()
-// 	os.Setenv("CLI_ENV_VAR", "Sound and fury")
-// 	os.Setenv("CLI_VAR", "As you like it")
-// 	checkDotEnv := false
-// 	want := noDotEnvExpectation
-// 	InitData(checkDotEnv, "example.env")
-// 	got, _ := Render("example.tmpl")
-// 	if got != want {
-// 		t.Fatalf("expected: %v, got: %v", want, got)
-// 	}
-// }